Make classic UPE checkout init resilient to deferred render-blocking JS
Host optimizers that defer all render-blocking JS (e.g. SiteGround Speed
Optimizer) relocate and defer our classic-checkout bundle without preserving
WordPress's script dependency order. Our module graph then evaluates before
window.wp.data / window.wp.i18n / window.wc.wcSettings exist; the first
externalized import read throws and the whole script aborts, so no card fields
render. 10.8.0's bundle slimming (native-globals polyfill mapping, #5475)
shifted timing enough to expose this; Apple Pay/Google Pay enables the
express-checkout bundle that tips the deferred order, and Pixel Manager was
coincidentally masking it by loading those globals early.

Split the classic UPE entry into a dependency-free bootstrap that waits for the
required globals, then loads the real init as an async chunk (so the externals
resolve only once they exist). The bootstrap is injected via JS, which a
defer-all optimizer does not reprocess. Pin webpack's publicPath to the plugin
build dir (localized inline, available even under deferral) so the async chunk
loads from the plugin rather than the optimizer's rewritten asset path.

Refs #5615
