+++ includes/payment-methods/class-wc-stripe-upe-payment-gateway.php
@@ -528,12 +528,18 @@ class WC_Stripe_UPE_Payment_Gateway extends WC_Stripe_Payment_Gateway {
 
 		$is_change_payment_method = $this->is_changing_payment_method_for_subscription();
 		$stripe_params            = [
-			'gatewayId'    => self::ID,
-			'title'        => $this->title,
-			'isUPEEnabled' => true,
-			'key'          => $this->publishable_key,
-			'locale'       => WC_Stripe_Helper::convert_wc_locale_to_stripe_locale( get_locale() ),
-			'apiVersion'   => WC_Stripe_API::STRIPE_API_VERSION,
+			'gatewayId'      => self::ID,
+			'title'          => $this->title,
+			'isUPEEnabled'   => true,
+			'key'            => $this->publishable_key,
+			'locale'         => WC_Stripe_Helper::convert_wc_locale_to_stripe_locale( get_locale() ),
+			'apiVersion'     => WC_Stripe_API::STRIPE_API_VERSION,
+			// Base URL for the classic UPE bundle's async init chunk. The JS
+			// bootstrap pins webpack's publicPath to this so the chunk loads from
+			// the plugin even when a host optimizer serves the entry script from a
+			// rewritten/relocated path. Localized inline (runs at parse time), so
+			// it is available before a deferred bundle executes.
+			'pluginBuildUrl' => trailingslashit( WC_STRIPE_PLUGIN_URL ) . 'build/',
 		];
 
 		$enabled_billing_fields = [];
